migrate to features architecture, fix leftover errors from userId to createdById changes.
