+++ src/app/(protected)/explore/page.tsx
@@ -1,5 +1,5 @@
 import { cookies } from 'next/headers';
-import { withAuth } from '~/lib/auth/protected';
+import { withAuth } from '~/auth/protected';
 
 export const metadata = {
   title: 'UCollab — Explore',
