--- src/app/user/[username]/page.tsx
@@ -1,146 +1,0 @@
-import { prisma } from '~/data/prisma';
-import { notFound } from 'next/navigation';
-import Image from 'next/image';
-
-export async function generateMetadata({ params }: { params: { username: string } }) {
-  const user = await prisma.user.findUnique({
-    where: { username: params.username },
-    select: { username: true },
-  });
-
-  if (!user) {
-    return { title: 'User Not Found' };
-  }
-
-  return { title: `${user.username} — Profile` };
-}
-
-interface UserProfile {
-  username: string;
-  email: string;
-  createdDate: string;
-  lastLogin: string;
-  verifiedEmail: boolean;
-  image: string | null;
-  profile: {
-    skills: string[];
-    gradYear: number | null;
-    bio: string | null;
-  } | null;
-}
-
-export default async function UserProfilePage({ params }: { params: { username: string } }) {
-  const { username } = params;
-
-  const user = await prisma.user.findUnique({
-    where: { username },
-    select: {
-      username: true,
-      email: true,
-      createdDate: true,
-      lastLogin: true,
-      avatar: true,
-      profile: {
-        select: {
-          skills: true,
-          gradYear: true,
-          bio: true,
-        },
-      },
-    },
-  });
-
-  if (!user) {
-    notFound();
-  }
-
-  const transformedUser: UserProfile = {
-    username: user.username,
-    email: user.email,
-    createdDate: user.createdDate.toISOString().split('T')[0],
-    lastLogin: user.lastLogin.toISOString().split('T')[0],
-    image: user.avatar,
-    profile: user.profile
-      ? {
-          skills: user.profile.skills ?? [],
-          gradYear: user.profile.gradYear,
-          bio: user.profile.bio,
-        }
-      : null,
-  };
-
-  return (
-    <>
-      <div className="absolute inset-0 flex h-full w-full flex-col items-center overflow-y-auto py-24">
-        <div className="flex w-full max-w-3xl items-center border-b p-4">
-          <Image
-            src={transformedUser.image ?? 'https://avatar.iran.liara.run/public'}
-            alt={transformedUser.username}
-            width={100}
-            height={100}
-            className="rounded-full"
-          />
-          <div className="ml-4">
-            <h1 className="flex items-center text-2xl font-bold">
-              {transformedUser.username}
-              {transformedUser.verifiedEmail ? (
-                <span className="ml-2 text-green-500">
-                  <svg xmlns="http://www.w3.org/2000/svg" width="1em" height="1em" viewBox="0 0 256 256">
-                    <path
-                      fill="currentColor"
-                      d="M225.86 102.82c-3.77-3.94-7.67-8-9.14-11.57c-1.36-3.27-1.44-8.69-1.52-13.94c-.15-9.76-.31-20.82-8-28.51s-18.75-7.85-28.51-8c-5.25-.08-10.67-.16-13.94-1.52c-3.56-1.47-7.63-5.37-11.57-9.14C146.28 23.51 138.44 16 128 16s-18.27 7.51-25.18 14.14c-3.94 3.77-8 7.67-11.57 9.14c-3.25 1.36-8.69 1.44-13.94 1.52c-9.76.15-20.82.31-28.51 8s-7.8 18.75-8 28.51c-.08 5.25-.16 10.67-1.52 13.94c-1.47 3.56-5.37 7.63-9.14 11.57C23.51 109.72 16 117.56 16 128s7.51 18.27 14.14 25.18c3.77 3.94 7.67 8 9.14 11.57c1.36 3.27 1.44 8.69 1.52 13.94c.15 9.76.31 20.82 8 28.51s18.75 7.85 28.51 8c5.25.08 10.67.16 13.94 1.52c3.56 1.47 7.63 5.37 11.57 9.14c6.9 6.63 14.74 14.14 25.18 14.14s18.27-7.51 25.18-14.14c3.94-3.77 8-7.67 11.57-9.14c3.27-1.36 8.69-1.44 13.94-1.52c9.76-.15 20.82-.31 28.51-8s7.85-18.75 8-28.51c.08-5.25.16-10.67 1.52-13.94c1.47-3.56 5.37-7.63 9.14-11.57c6.63-6.9 14.14-14.74 14.14-25.18s-7.51-18.27-14.14-25.18m-11.55 39.29c-4.79 5-9.75 10.17-12.38 16.52c-2.52 6.1-2.63 13.07-2.73 19.82c-.1 7-.21 14.33-3.32 17.43s-10.39 3.22-17.43 3.32c-6.75.1-13.72.21-19.82 2.73c-6.35 2.63-11.52 7.59-16.52 12.38S132 224 128 224s-9.15-4.92-14.11-9.69s-10.17-9.75-16.52-12.38c-6.1-2.52-13.07-2.63-19.82-2.73c-7-.1-14.33-.21-17.43-3.32s-3.22-10.39-3.32-17.43c-.1-6.75-.21-13.72-2.73-19.82c-2.63-6.35-7.59-11.52-12.38-16.52S32 132 32 128s4.92-9.15 9.69-14.11s9.75-10.17 12.38-16.52c2.52-6.1 2.63-13.07 2.73-19.82c.1-7 .21-14.33 3.32-17.43s10.39-3.22 17.43-3.32c6.75-.1 13.72-.21 19.82-2.73c6.35-2.63 11.52-7.59 16.52-12.38S124 32 128 32s9.15 4.92 14.11 9.69s10.17 9.75 16.52 12.38c6.1 2.52 13.07 2.63 19.82 2.73c7 .1 14.33.21 17.43 3.32s3.22 10.39 3.32 17.43c.1 6.75.21 13.72 2.73 19.82c2.63 6.35 7.59 11.52 12.38 16.52S224 124 224 128s-4.92 9.15-9.69 14.11"
-                    />
-                  </svg>
-                </span>
-              ) : (
-                <span className="mt-1 ml-2 text-orange-500">
-                  <svg xmlns="http://www.w3.org/2000/svg" width="1em" height="1em" viewBox="0 0 256 256">
-                    <path
-                      fill="currentColor"
-                      d="M225.86 102.82c-3.77-3.94-7.67-8-9.14-11.57c-1.36-3.27-1.44-8.69-1.52-13.94c-.15-9.76-.31-20.82-8-28.51s-18.75-7.85-28.51-8c-5.25-.08-10.67-.16-13.94-1.52c-3.56-1.47-7.63-5.37-11.57-9.14C146.28 23.51 138.44 16 128 16s-18.27 7.51-25.18 14.14c-3.94 3.77-8 7.67-11.57 9.14c-3.25 1.36-8.69 1.44-13.94 1.52c-9.76.15-20.82.31-28.51 8s-7.8 18.75-8 28.51c-.08 5.25-.16 10.67-1.52 13.94c-1.47 3.56-5.37 7.63-9.14 11.57C23.51 109.72 16 117.56 16 128s7.51 18.27 14.14 25.18c3.77 3.94 7.67 8 9.14 11.57c1.36 3.27 1.44 8.69 1.52 13.94c.15 9.76.31 20.82 8 28.51s18.75 7.85 28.51 8c5.25.08 10.67.16 13.94 1.52c3.56 1.47 7.63 5.37 11.57 9.14c6.9 6.63 14.74 14.14 25.18 14.14s18.27-7.51 25.18-14.14c3.94-3.77 8-7.67 11.57-9.14c3.27-1.36 8.69-1.44 13.94-1.52c9.76-.15 20.82-.31 28.51-8s7.85-18.75 8-28.51c.08-5.25.16-10.67 1.52-13.94c1.47-3.56 5.37-7.63 9.14-11.57c6.63-6.9 14.14-14.74 14.14-25.18s-7.51-18.27-14.14-25.18m-11.55 39.29c-4.79 5-9.75 10.17-12.38 16.52c-2.52 6.1-2.63 13.07-2.73 19.82c-.1 7-.21 14.33-3.32 17.43s-10.39 3.22-17.43 3.32c-6.75.1-13.72.21-19.82 2.73c-6.35 2.63-11.52 7.59-16.52 12.38S132 224 128 224s-9.15-4.92-14.11-9.69s-10.17-9.75-16.52-12.38c-6.1-2.52-13.07-2.63-19.82-2.73c-7-.1-14.33-.21-17.43-3.32s-3.22-10.39-3.32-17.43c-.1-6.75-.21-13.72-2.73-19.82c-2.63-6.35-7.59-11.52-12.38-16.52S32 132 32 128s4.92-9.15 9.69-14.11s9.75-10.17 12.38-16.52c2.52-6.1 2.63-13.07 2.73-19.82c.1-7 .21-14.33 3.32-17.43s10.39-3.22 17.43-3.32c6.75-.1 13.72-.21 19.82-2.73c6.35-2.63 11.52-7.59 16.52-12.38S124 32 128 32s9.15 4.92 14.11 9.69s10.17 9.75 16.52 12.38c6.1 2.52 13.07 2.63 19.82 2.73c7 .1 14.33.21 17.43 3.32s3.22 10.39 3.32 17.43c.1 6.75.21 13.72 2.73 19.82c2.63 6.35 7.59 11.52 12.38 16.52S224 124 224 128s-4.92 9.15-9.69 14.11M120 136V80a8 8 0 0 1 16 0v56a8 8 0 0 1-16 0m20 36a12 12 0 1 1-12-12a12 12 0 0 1 12 12"
-                    />
-                  </svg>
-                </span>
-              )}
-            </h1>
-            <p className="text-accent">{transformedUser.email}</p>
-            <div className="text-accent flex items-center text-sm">
-              <svg xmlns="http://www.w3.org/2000/svg" width="1em" height="1em" viewBox="0 0 256 256" className="mr-1">
-                <path
-                  fill="currentColor"
-                  d="M208 32h-24v-8a8 8 0 0 0-16 0v8H88v-8a8 8 0 0 0-16 0v8H48a16 16 0 0 0-16 16v160a16 16 0 0 0 16 16h160a16 16 0 0 0 16-16V48a16 16 0 0 0-16-16M72 48v8a8 8 0 0 0 16 0v-8h80v8a8 8 0 0 0 16 0v-8h24v32H48V48Zm136 160H48V96h160zm-68-76a12 12 0 1 1-12-12a12 12 0 0 1 12 12m44 0a12 12 0 1 1-12-12a12 12 0 0 1 12 12m-88 40a12 12 0 1 1-12-12a12 12 0 0 1 12 12m44 0a12 12 0 1 1-12-12a12 12 0 0 1 12 12m44 0a12 12 0 1 1-12-12a12 12 0 0 1 12 12"
-                />
-              </svg>
-              <span>Joined {transformedUser.createdDate}</span>
-            </div>
-            {transformedUser.profile?.gradYear && (
-              <div className="text-accent flex items-center text-sm">
-                <svg xmlns="http://www.w3.org/2000/svg" width="1em" height="1em" viewBox="0 0 256 256" className="mr-1">
-                  <path
-                    fill="currentColor"
-                    d="m251.76 88.94l-120-64a8 8 0 0 0-7.52 0l-120 64a8 8 0 0 0 0 14.12L32 117.87v48.42a15.9 15.9 0 0 0 4.06 10.65C49.16 191.53 78.51 216 128 216a130 130 0 0 0 48-8.76V240a8 8 0 0 0 16 0v-40.49a115.6 115.6 0 0 0 27.94-22.57a15.9 15.9 0 0 0 4.06-10.65v-48.42l27.76-14.81a8 8 0 0 0 0-14.12M128 200c-43.27 0-68.72-21.14-80-33.71V126.4l76.24 40.66a8 8 0 0 0 7.52 0L176 143.47v46.34c-12.6 5.88-28.48 10.19-48 10.19m80-33.75a97.8 97.8 0 0 1-16 14.25v-45.57l16-8.53Zm-20-47.31l-.22-.13l-56-29.87a8 8 0 0 0-7.52 14.12L171 128l-43 22.93L25 96l103-54.93L231 96Z"
-                  />
-                </svg>
-                <span>Graduating in {transformedUser.profile.gradYear}</span>
-              </div>
-            )}
-          </div>
-        </div>
-        <div className="mt-2 w-full max-w-3xl px-4">
-          <span className="font-semibold">Biography:</span>
-          <div className="mt-2">{transformedUser.profile?.bio ?? 'No bio available'}</div>
-          <div className="mt-2">
-            <span className="font-semibold">Skills:</span>
-            <ul className="list-inside list-disc">
-              {transformedUser.profile?.skills && transformedUser.profile.skills.length ? (
-                transformedUser.profile.skills.map((skill, index) => <li key={index}>{skill}</li>)
-              ) : (
-                <li>No skills listed</li>
-              )}
-            </ul>
-          </div>
-        </div>
-      </div>
-    </>
-  );
-}
